Define terms: bullet chatting and comment
It is good practice in technical specs to define terms and refer to them
throughout the spec. That usually helps understanding and exposing links between
concepts.

In particular, separating "bullet chatting" (the entire experience) and
"bullet chatting comment" (an individual comment) allows to distinguish between
properties that apply to the entire experience and those that apply to each
comment.

Also, it does not make sense to define WebVTT and TTML, which simply need to
appear as references.

This update modifies the Terminology section accordingly, reformulates the
Introduction, and moves figures about commercial operation to an appendix.